py-from-zero-to-hero-06/requirements.txt
# For Windows
waitress==3.0.2
chromadb==0.6.3
elevenlabs==1.55.0
fastembed==0.6.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.49
langchain-ollama==0.3.0
langchain-text-splitters==0.3.7
ollama==0.4.7
unstructured==0.17.2
unstructured[all-docs]
redis==5.2.1
sentence-transformers==4.0.1
numpy>=1.22.4,<2.0.0